+++ src/enums/ViewMethodEnum.ts
@@ -3,6 +3,7 @@ enum ViewMethodEnum {
   GetVersion = 'get_version',
   IsWritePermissionGranted = 'is_write_permission_granted',
   StorageBalanceOf = 'storage_balance_of',
+  Keys = 'keys',
 }
 
 export default ViewMethodEnum;