+++ src/errors/index.ts
@@ -2,3 +2,4 @@ export { default as AccountNotFoundError } from './AccountNotFoundError';
 export { default as BaseError } from './BaseError';
 export { default as InvalidAccountIdError } from './InvalidAccountIdError';
 export { default as KeyNotAllowedError } from './KeyNotAllowedError';
+export { default as UnknownNetworkError } from './UnknownNetworkError';