+++ src/types/IAccount.ts
@@ -1,0 +1,8 @@
+import { utils } from 'near-api-js';
+
+interface IAccount {
+  accountID: string;
+  publicKey: string | utils.PublicKey;
+}
+
+export default IAccount;