+++ src/types/IDefaultOptions.ts
@@ -1,7 +1,0 @@
-import type { Account } from 'near-api-js';
-
-interface IDefaultOptions {
-  signer: Account;
-}
-
-export default IDefaultOptions;