+++ src/types/IDefaultViewOptions.ts
@@ -1,6 +1,0 @@
-// types
-import type IDefaultOptions from './IDefaultOptions';
-
-type IDefaultViewOptions = IDefaultOptions;
-
-export default IDefaultViewOptions;