+++ src/types/IGetVersionOptions.ts
@@ -1,5 +1,0 @@
-import type IDefaultViewOptions from './IDefaultViewOptions';
-
-type IGetVersionOptions = IDefaultViewOptions;
-
-export default IGetVersionOptions;