+++ src/types/INewSocialOptions.ts
@@ -1,5 +1,10 @@
+// types
+import type IRPCOptions from './IRPCOptions';
+
 interface INewSocialOptions {
   contractId?: string;
+  network?: string | IRPCOptions;
+  apiServer?: string;
 }
 
 export default INewSocialOptions;