+++ src/utils/rpcQueries/index.ts
@@ -1,0 +1,2 @@
+export * from './viewAccessKeyList';
+export * from './viewFunction';