+++ src/utils/rpcQueries/viewAccessKeyList/types/IAccessKeyInfoViewRaw.ts
@@ -1,0 +1,8 @@
+import type { AccessKeyViewRaw } from '@near-js/types';
+
+interface IAccessKeyInfoViewRaw {
+  access_key: AccessKeyViewRaw;
+  public_key: string;
+}
+
+export default IAccessKeyInfoViewRaw;